Adds on-board flash filesystem support
Enables support for the on-board flash filesystem on the IoTNode LoRaWAN board.

This change adds the ability to utilize the flash memory as a filesystem,
allowing for storage of configuration files, sensor data, or other
persistent data.

Includes modifications to the build process to allocate space for
filesystem, as well as code to initialize and check the filesystem's
availability. Also, displays file system information in about output.